backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Type hints and validation
pydantic==2.10.5
pydantic-settings==2.7.1

# CORS support
python-multipart==0.0.20

# Git integration
GitPython==3.1.43
